// apps/api/src/generated/prisma/enums.ts
export const VerificationDecisionAction = {
  NEEDS_INFORMATION: 'NEEDS_INFORMATION',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED'
} as const

export type VerificationDecisionAction = (typeof VerificationDecisionAction)[keyof typeof VerificationDecisionAction]